microservices/user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Ivan Gruzdev <gruzdev.ivan.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.65.1"
grpcio-tools = "^1.65.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
pydantic = {extras = ["email"], version = "^2.8.2"}
alembic = "^1.13.2"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
